// src/pages/articles.module.scss.d.ts
// generated with typed-scss-modules
export type Styles = {
  'cardLabel': string;
  'date': string;
  'grid': string;
  'qiita': string;
  'tag': string;
  'tagContainer': string;
  'card': string;
  'cardContainer': string;
  'label': string;
};

export type TheClasses = keyof Styles;

declare const styles: Styles;

export default styles;